Move PairingManager to imbot/security (reusable infrastructure)
PairingManager (TOFU code generation, constant-time verify, lockout) is a
pure mechanism with no remote-control dependencies. Moving it to
imbot/security/ makes it available to any application built on the imbot
framework, consistent with imbot's role as standalone infrastructure.

Changes:
- imbot/security/pairing.go: PairingAuditor interface + full PairingManager
  (audit dependency replaced by PairingAuditor interface; *audit.Logger
  satisfies it automatically via Go structural typing)
- imbot/security/pairing_test.go: tests moved from remote_control/bot/
- internal/remote_control/bot/pairing_manager.go: replaced with thin alias
  layer (type aliases + var aliases) — all existing callers unchanged
- internal/remote_control/bot/pairing_manager_test.go: deleted (tests live
  in imbot/security/)
- internal/remote_control/bot/bot_handler_pair.go: removed stale audit
  compile-time guard
