+++ imbot/security/pairing_test.go
@@ -1,4 +1,4 @@
-package bot
+package security
 
 import (
 	"errors"
